requirements.txt
pageindex>=0.1.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
streamlit>=1.35.0
pandas>=2.2.0
python-multipart>=0.0.9
pytest>=8.0.0
